Move methods from presetshandler to cameracontroller
